feat(nuxt): allow opting out of payload route restoration
