+++ docs/3.guide/6.going-further/1.experimental-features.md
@@ -158,6 +158,20 @@ export default defineNuxtConfig({
 })
 ```
 
+## restorePayloadRoute
+
+Controls whether Nuxt uses the server-rendered payload route during hydration when it differs from the browser URL.
+
+This is enabled by default to avoid hydration mismatches when a prerendered or cached page is loaded from another URL. You can disable it if your deployment intentionally serves the same Nuxt response from multiple external URLs and the visible browser URL must be preserved.
+
+```ts twoslash [nuxt.config.ts]
+export default defineNuxtConfig({
+  experimental: {
+    restorePayloadRoute: false,
+  },
+})
+```
+
 ## inlineRouteRules
 
 Define route rules at the page level using [`defineRouteRules`](/docs/4.x/api/utils/define-route-rules).
