+++ packages/nuxt/src/core/templates.ts
@@ -555,6 +555,7 @@ export const nuxtConfigTemplate: NuxtTemplate = {
       `export const componentIslands = ${shouldEnableComponentIslands}`,
       `export const componentIslandsActive = ${componentIslandsActive}`,
       `export const payloadExtraction = ${payloadExtraction}`,
+      `export const restorePayloadRoute = ${!!ctx.nuxt.options.experimental.restorePayloadRoute}`,
       `export const prefetchPreloadTags = ${!!ctx.nuxt.options.experimental.prefetchPreloadTags}`,
       `export const cookieStore = ${!!ctx.nuxt.options.experimental.cookieStore}`,
       `export const appManifest = ${!!ctx.nuxt.options.experimental.appManifest}`,
